rocko: bring in graphics and multimedia from meta-freescale next
meta-freescale only supports the linux kernel 4.1 based bsp drivers.
Since we are using the latest 4.9 drivers we need to pull in some
of the upgraded recipes to support our kernel.  This mainline grabs
the graphics libraries, multimedia librariers, and firmware files
needed to build and run rocko on the SolidRun iMX6 based boards.

Signed-off-by: Jon Nettleton <jon@solid-run.com>