Bring in the meta-freescale qt5 dynamic layer with patches
This adds the QT5 dynamic layer to our meta-layer.  This fixes
builds in rocko for the eglfs-viv and eglfs-viv-wl drivers. I have
also changed the default QT build to properly default to the correct
drivers depending what graphics framework is chosen.  The order of
preference is X11, Wayland, Framebuffer.  The decision is made
based on DISTRO_FEATURES.

Signed-off-by: Jon Nettleton <jon@solid-run.com>